+++ run.sh
@@ -2,7 +2,13 @@
 
 set -euo pipefail
 
+outfile="output/avery_5167.pdf"
+
+rm -f $outfile
+
 /opt/homebrew/opt/python@3.12/bin/python3.12 lbx_to_avery_5167.py \
 	LEGO_BRICK_LABELS-v40/Labels \
-	--output output/avery_5167.pdf \
+	--output $outfile \
 	--draw-outlines
+
+echo "output written to: $outfile"